+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # What's New
 
+<a name="unreleased"></a>
+## [Unreleased]
+
+### Improvements/Bug Fixes
+- inherit AnlzFile from Mapping to implement dict interface
+- unify binary file API
+
+
 <a name="0.0.1"></a>
 ## [0.0.1] - 2022-10-04
-
 ### Improvements/Bug Fixes
 - fix Python version
 
@@ -11,4 +18,5 @@
 ## 0.0.0 - 2022-10-04
 
 
+[Unreleased]: https://github.com/dylanljones/pyrekordbox/compare/0.0.1...HEAD
 [0.0.1]: https://github.com/dylanljones/pyrekordbox/compare/0.0.0...0.0.1